+++ src/main/java/com/financialtargets/incomes/application/utils/AmountUtil.java
@@ -7,10 +7,10 @@ import java.math.RoundingMode;
 
 @UtilityClass
 public class AmountUtil {
-    public String formatAmount(Float amount) {
+    public String formatAmount(BigDecimal amount) {
         if (amount == null) return "R$ 0,00";
 
-        BigDecimal decimal = BigDecimal.valueOf(amount).setScale(2, RoundingMode.HALF_UP);
+        BigDecimal decimal = amount.setScale(2, RoundingMode.HALF_UP);
 
         return "R$ " + decimal.toString().replace(".", ",");
     }
